% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.distances.R
\name{get.distances}
\alias{get.distances}
\title{The distances between the data, neurons, and across both .}
\usage{
get.distances(map)
}
\arguments{
\item{map}{map; object returned by map.build}
}
\value{
List containing distances between 3 matrices:
dist.data - distance between all data points
dist.neurons - distance between all map neurons
dist.cross - distance between the data points and the map neurons
dist.proj - distance between the projected data points
}
\description{
The distances between the data, neurons, and across both .
}
