% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.hood.R
\name{get.hood}
\alias{get.hood}
\title{Get the indices of the neighboring nodes}
\usage{
get.hood(x, y, xdim, ydim)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{xdim}{TODO}

\item{ydim}{TODO}
}
\value{
TODO
}
\description{
Get the indices of the neighboring nodes
}
