% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.map.coord.R
\name{get.map.coord}
\alias{get.map.coord}
\title{Gets the x, y position in the map}
\usage{
get.map.coord(i, xdim, ydim)
}
\arguments{
\item{i}{TODO}

\item{xdim}{TODO}

\item{ydim}{TODO}
}
\value{
TODO
}
\description{
Gets the x, y position in the map
}
