% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.map.diff.R
\name{get.map.diff}
\alias{get.map.diff}
\title{The difference between two maps , i.e. the average distance between the closest pairs of neurons .}
\usage{
get.map.diff(map1, map2)
}
\arguments{
\item{map1}{TODO}

\item{map2}{TODO}
}
\value{
TODO:
}
\description{
The difference between two maps , i.e. the average distance between the closest pairs of neurons .
}
