% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.map.idx.R
\name{get.map.idx}
\alias{get.map.idx}
\title{Get the index in the map}
\usage{
get.map.idx(x, y, xdim)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{xdim}{TODO}
}
\value{
TODO
}
\description{
Get the index in the map
}
