% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.label.accuracy.R
\name{map.label.accuracy}
\alias{map.label.accuracy}
\title{The labeling of the neurons and other measures}
\usage{
map.label.accuracy(map)
}
\arguments{
\item{map}{map;an object returned by map. build}
}
\value{
list containing :
neurons - majority label mapped to each neuron
acc - labeling accuracy of the map
bmu - ratio of BMUs
fit - combined accuracy and ratio value
}
\description{
The labeling of the neurons and other measures
}
