% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot2d.R
\name{map.plot2d}
\alias{map.plot2d}
\title{Plot the map in 2 - dimensions with data point and neurons colored by label}
\usage{
map.plot2d(
  map,
  x.idx = 1,
  y.idx = 2,
  show.dead = F,
  highlight = NULL,
  use.rgl = F
)
}
\arguments{
\item{map}{TODO}

\item{x.idx}{TODO}

\item{y.idx}{TODO}

\item{show.dead}{TODO}

\item{highlight}{TODO}

\item{use.rgl}{TODO}
}
\value{
TODO
}
\description{
Plot the map in 2 - dimensions with data point and neurons colored by label
}
