% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot3d.R
\name{map.plot3d}
\alias{map.plot3d}
\title{Plot the map in 3 - dimensions with data point and neurons colored by label}
\usage{
map.plot3d(
  map,
  x.idx = 1,
  y.idx = 2,
  z.idx = 3,
  show.data = T,
  show.map = T,
  show.dead = F,
  highlight = NULL,
  use.rgl = F
)
}
\arguments{
\item{map}{TODO}

\item{x.idx}{TODO}

\item{y.idx}{TODO}

\item{z.idx}{TODO}

\item{show.dead}{TODO}

\item{highlight}{TODO}

\item{use.rgl}{TODO}
}
\value{
TODO
}
\description{
Plot the map in 3 - dimensions with data point and neurons colored by label
}
